only rename file cmake-tools.ts
