+++ src/drivers/cmfileapi-driver.ts
@@ -1,5 +1,5 @@
 import * as api from '@cmt/api';
-import { ConfigureTrigger } from '@cmt/cmake-tools';
+import { ConfigureTrigger } from '@cmt/cmakeTools';
 import { ExecutableTarget } from '@cmt/api';
 import { CMakeCache } from '@cmt/cache';
 import { CMakeExecutable } from '@cmt/cmake/cmake-executable';
