+++ test/extension-tests/multi-root-UI/test/debugger.test.ts
@@ -1,6 +1,6 @@
 import { DefaultEnvironment, expect, getFirstSystemKit } from '@test/util';
 import * as vscode from 'vscode';
-import CMakeTools from '@cmt/cmake-tools';
+import CMakeTools from '@cmt/cmakeTools';
 
 suite('Debug/Launch interface', async () => {
     let testEnv: DefaultEnvironment;
