+++ test/extension-tests/single-root-UI/test/configure-and-build-kits.test.ts
@@ -9,7 +9,7 @@ import {
 } from '@test/util';
 import * as path from 'path';
 import * as vscode from 'vscode';
-import CMakeTools from '@cmt/cmake-tools';
+import CMakeTools from '@cmt/cmakeTools';
 
 const workername: string = process.platform;
