+++ test/extension-tests/single-root-UI/test/debugger-kits.test.ts
@@ -1,6 +1,6 @@
 import { DefaultEnvironment, expect, getFirstSystemKit } from '@test/util';
 import * as vscode from 'vscode';
-import CMakeTools from '@cmt/cmake-tools';
+import CMakeTools from '@cmt/cmakeTools';
 
 suite('Debug/Launch interface using Kits and Variants', async () => {
     let testEnv: DefaultEnvironment;
