+++ test/extension-tests/successful-build/test/toolchain.test.ts
@@ -1,7 +1,7 @@
 /* eslint-disable no-unused-expressions */
 import * as api from '@cmt/api';
 import { CMakeCache } from '@cmt/cache';
-import { CMakeTools, ConfigureTrigger } from '@cmt/cmake-tools';
+import { CMakeTools, ConfigureTrigger } from '@cmt/cmakeTools';
 import { readKitsFile, kitsForWorkspaceDirectory, getAdditionalKits, USER_KITS_FILEPATH } from '@cmt/kit';
 import { platformNormalizePath } from '@cmt/util';
 import { DefaultEnvironment, expect } from '@test/util';
