+++ test/extension-tests/successful-build/test/variable-substitution.test.ts
@@ -1,6 +1,6 @@
 import * as api from '@cmt/api';
 import { CMakeCache } from '@cmt/cache';
-import { CMakeTools, ConfigureTrigger } from '@cmt/cmake-tools';
+import { CMakeTools, ConfigureTrigger } from '@cmt/cmakeTools';
 import paths from '@cmt/paths';
 import { objectPairs, platformNormalizePath, makeHashString } from '@cmt/util';
 import { clearExistingKitConfigurationFile, DefaultEnvironment, expect, getFirstSystemKit } from '@test/util';
